Fixed some errors

+++ util/util.js
@@ -13,15 +13,18 @@ module.exports.print = print;
 function clearSessions() {
 	const directory = 'sessions';
 
-	fs.readdir(directory, (err, files) => {
-	  if (err) throw err;
+	if (fs.existsSync(directory)) {
+		fs.readdir(directory, (err, files) => {
+			if (err) throw err;
 
-	  for (const file of files) {
-	    fs.unlink(path.join(directory, file), err => {
-	      if (err) throw err;
-	    });
-	  }
-	});
+			for (const file of files) {
+				fs.unlink(path.join(directory, file), err => {
+					if (err) throw err;
+				});
+			}
+		});
+	}
+
 }
 module.exports.clearSessions = clearSessions;
